feat: 完善Dockerfile文件和一些maven编译时扫描到的规范问题的修复

+++ peach-middleware/peach-openfeign/peach-openfeign-autoconfigure/src/main/java/com/peach/openfeign/support/PeachOpenFeignErrorDecoder.java
@@ -71,7 +71,7 @@ public class PeachOpenFeignErrorDecoder implements ErrorDecoder {
 
         PeachFeignRetryableException(int status, String message, Request.HttpMethod httpMethod, Request request,
                                      Throwable cause) {
-            super(status, message, httpMethod, cause, null, request);
+            super(status, message, httpMethod, cause, (Long)null, request);
         }
     }
